GAP-130: UUID v7 identifiers, written as strings, shown as tags; /v3 routes (#130)
* GAP-130: a journal written before the identifier change

DN-31 §9 row 1 (D-56): a pre-change journal replays and reports
unchanged. A desktop session written at 7d6f201 by a temporary
gungnir-app test, run once and deleted, because after GAP-130 the same
code mints UUID v7 identifiers; its plans 1 to 3 and decision 1 are the
old u64 counters. It holds PlanProposed, Decided, an engagement opened
and closed, a handoff issued and reported, an escalation and an expiry;
report.json is JournalReportGenerator's export of it. SOURCE.md says how,
and which line stands in for the node's record. .gitignore admits the
journal as a fixture.


* GAP-130: identifiers minted as UUID v7, written as strings, shown as tags

D-56, D-60, D-61; DN-31 §5.1 amendment 1. DecisionId, PlanId and
PendingApprovalId hold a u128 minted with Uuid::now_v7 in the workflow and
the planner, replacing counters that restarted at 1 per process.
gungnir_model::identifier writes the hyphenated form and reads it or a
number, so old journals read; the handoff delivered to an effector and
published for exchange are no longer null. Panels and alerts show the short
tag, PN-07 and PN-20 the whole id with a copy control, logs and audit
entries the whole id. SCHEMA_VERSION is 4 and ApprovalRequested is removed.
The rationale names no throwaway plan id, and the delivery stub now checks
the body it receives.


* GAP-130: routes move to /v3 and every /v2 route answers 410 Gone

D-56, DN-31 §5.1 and §7. The identifier change altered every payload
carrying a plan or a decision, so the interface moves whole. gungnir_api
builds the node's routes and gungnir-remote's URLs from one constant,
API_VERSION, through path() and routes. The v2 module is v3. Each /v2 route
authenticates its caller as its /v3 successor does, then answers 410 with
the successor in the problem; the /v2 event stream answers before any
upgrade. The report route reads a decision as the hyphenated UUID or a
decimal number and refuses other text with a readable 400. The /v3 decision
route still refuses with 501 until GAP-132. retired_v2.rs holds all of it.


* GAP-130: DN-31 §9 row 1's tests, and plans across an outage

DN-31 §9 row 1, D-56. Separate workflows and planners standing in for a
node, two desktops and a restart mint no equal identifier, each a v7 in
creation order. With two desktops' handoffs present, a report posted to
the node reaches only its own through the node's queue and each link
inbox. MOE-05 over two machines' merged journals pairs each engagement
with its own decision. The pre-change journal replays and regenerates its
committed report byte for byte, and rewrites without loss. The failover
test walks an outage: the counters hid the node's plan behind the embedded
planner's plan 1 at the switch back.


* GAP-130: identifier serde impls written out, declarations kept plain

D-60. With #[serde(with)] inside the tuple, the declarations no longer
matched the shape the UAF information model is read in, and DecisionId
and PlanId dropped out of If-Sr with every relationship to them. The
Serialize and Deserialize impls now call gungnir_model::identifier::wire
on the type, so each declaration reads as DN-31 §5.1 shows it:
pub struct DecisionId(pub u128). The written form is unchanged.


* GAP-130: DN-31 amendment 1, D-60 and D-61, GAP-135, and the record

DN-31 amendment 1 records what the build found: a 128-bit JSON number
breaks to_value, Value parsing and field-tagged enums, and would have
nulled the handoff delivery and exchange bodies. D-60 writes identifiers
as strings and D-61 shows them as tags; the §9 citation is corrected and
ApprovalRequested's removal placed here. GAP-130 closes; GAP-135 files the
engagement a report moves without a record. The API contract, the
architecture tables, the UX map and the schema deltas describe /v3 and
SCHEMA_VERSION 4, and the UAF is regenerated with its four changed
information diagrams re-rendered. The record item lists every minting site
and order assumption.


* GAP-130: an effector report's audit entry names the decision whole

D-61: audit entries carry the whole identifier and alerts the short tag.
apply_report built one sentence for both, so the audit entry carried the
tag; the sentence now leaves the decision out, and each names it in its own
form. The PlanId and fresh_plan comments said the counter collision bit at
the fall back; it bit at the switch back, as the failover test and the
record item say.


---------

+++ gungnir-app/tests/authentication.rs
@@ -493,7 +493,7 @@ fn an_expired_session_returns_the_desktop_to_the_selected_role() {
 }
 
 /// Queue a plan directly: the desktop has no tracks, so it cannot produce one.
-fn queue(state: &mut AppState, plan: u64) -> gungnir_command::PendingApprovalId {
+fn queue(state: &mut AppState, plan: u128) -> gungnir_command::PendingApprovalId {
     use gungnir_command::{ApprovalWorkflow, Submission};
     state
         .approvals
